fix: fix multiple issues and bugs for stable release
Fixed:
- Create orders with more than one food (bug)
- Find by date not working
- Food struct changed due in one petition it updated the null values
- Bug to buy less than one kilogram
- Get money with non-integer units

+++ pkg/features/money/struct/money.struct.go
@@ -1,7 +1,7 @@
 package moneystruct
 
 type CreateMoney struct {
-	Amount      int     `validate:"required"`
+	Amount      float64 `validate:"required"`
 	Reason      string  `validate:"required"`
 	Description *string `validate:"omitempty"`
 }
